pyproject.toml
[project]
name = "apptitude"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "black>=25.9.0",
    "fastapi[standard]>=0.116.1",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.27",
    "langchain-cohere>=0.4.5",
    "langchain-google-genai>=2.0.10",
    "langchain-postgres>=0.0.15",
    "langchain-text-splitters>=0.3.10",
    "openai>=1.102.0",
    "pdf2image>=1.17.0",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "ruff>=0.13.2",
    "selenium>=4.36.0",
    "uvicorn>=0.35.0",
    "werkzeug>=3.1.3",
]
